+++ src/main/java/com/prography/lighton/performance/presentation/dto/PerformanceUpdateRequest.java
@@ -5,12 +5,13 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
 import org.hibernate.validator.constraints.URL;
 
 public record PerformanceUpdateRequest(
+
+        @NotEmpty(message = "아티스트는 하나 이상 선택해야 합니다.")
+        List<@NotNull(message = "아티스트 ID는 필수입니다.") Long> artists,
 
         @NotNull(message = "공연 기본 정보는 필수입니다.")
         @Valid
@@ -32,62 +33,4 @@ public record PerformanceUpdateRequest(
         String poster
 
 ) {
-    public record Info(
-
-            @NotEmpty(message = "아티스트는 하나 이상 선택해야 합니다.")
-            List<@NotNull(message = "아티스트 ID는 필수입니다.") Long> artists,
-
-            @NotBlank(message = "공연명은 필수입니다.")
-            String title,
-
-            @NotBlank(message = "공연 소개는 필수입니다.")
-            String description,
-
-            @NotNull(message = "지역 코드는 필수입니다.")
-            Integer location,
-
-            @NotBlank(message = "공연 장소는 필수입니다.")
-            String place,
-
-            String notice,
-
-            @NotEmpty(message = "장르는 하나 이상 선택해야 합니다.")
-            List<@NotNull(message = "장르 ID는 필수입니다.") Long> genre
-
-    ) {
-    }
-
-    public record Schedule(
-
-            @NotNull(message = "공연 시작 날짜는 필수입니다.")
-            LocalDate startDate,
-
-            @NotNull(message = "공연 종료 날짜는 필수입니다.")
-            LocalDate endDate,
-
-            @NotNull(message = "공연 시작 시간은 필수입니다.")
-            LocalTime startTime,
-
-            @NotNull(message = "공연 종료 시간은 필수입니다.")
-            LocalTime endTime
-
-    ) {
-    }
-
-    public record Payment(
-
-            @NotNull(message = "비용 유뮤는 필수입니다.")
-            Boolean isPaid,
-
-            Integer price,
-
-            String account,
-
-            String bank,
-
-            String accountHolder
-
-    ) {
-    }
 }
-
